Abandoning scroll based approach, it's jittery garbage.
New plan: Attempt to create a clone of the target append it to the
body. Strip all positioning and sizing styles. Match up per scroll/frame to the
X and dimensions of the real element.

No wait, that won't work because the element may have inner elements that are
effecting the sticky'd element.

Can we thrash per scroll fast enough to get away with it?

Position fixing is what's causing the problem and takes the element out of the flow.
Is the solution to clone the container and swap its attachment based on scroll?

If the parent's height is auto it's not really sticky and there won't be sticky
taking effect.

If the parent's height is width auto it will need to be set by the element to be stickied.
So moving the element to a cloned container would break the original container's styles.

jCan the innerHTML on the clone be updated fast enough?

We will have to intentionally thrash if the window or container resize.
Track all applied transforms and reverse them if/when necessary.
Use translate2d to do matrix based updates.
Swap in and out of fixed.

There will be a need for a lot of value tracking and it would probably
be best to just restart from the original state of Sticky.
When in the middle, blow out margins, top/left and adjust translations.

When container or window resize, clear set inline styles and re-measure.

Padding, width and height will need to be measured pre-fixing and set
with inline styles.

We'll need to make note of a caveat that any element using Sticky should
not be subjected to other effects that might apply inline styles.